// src/app/pages/dashboard/learn/learn.interface.ts
/*
 * File: learn.interface.ts
 * Project: staysafe-frontend
 * File Created: Saturday, 11th April 2020 8:13:33 pm
 * Author: Adithya Sreyaj
 * -----
 * Last Modified: Saturday, 11th April 2020 8:13:48 pm
 * Modified By: Adithya Sreyaj<adi.sreyaj@gmail.com>
 * -----
 */

export interface LearnLinks {
  title: string;
  link: string;
}